Simplify secrets management for the repo
Uses my standard approach to managing secrets in a repo

Adds the `secerts` script for easily encrypting and decrypting
secrets stored in the `secrets` directory as part of the repo.
Also add some nice ergonomic features with `.envrc` and `.gitingore`
to assure `secrets` is in your path and ignore unencrypted secrets.