Initial lab deploy (#2)
TL;DR
-----

Provides for a TKG lab installation with some personal additions

Details
-------

Allows me to set up a lab using the TKG Lab codebase with some added
capabilities to tailor the lab for my purposes. Uses my standard 
model for managing secrets with a simple Python script and storing
the PGP encrypted secrets in the `secrets` directory. Also uses
the `.envrc` to provide some simple ergonomics.

The expected `PARAMS_YAML` for TKG Lab is stored as a secret. It's
current with the TKG 1.4 version of the lab as of today, and has 
some additional values that are useful for my changes and 
additions to the lab. At this point I have not documented each of 
my changes into a non-encrypted file, but I plan to do that when I
write up some docs on how to work with the repo. I can use my file
directly with TKG Lab as is. I also made the file into a YTT data 
values file, since that made my customizations easier.

Customizations
--------------

I've customized the lab in two primary ways. I've replaced 
Kubernetes secrets for Concourse with Vault, and I've added 
Minio as a blob store for Harbor. I've mad a smaller tweak to
certificate manager to get around issues I was having with my 
split-horizon DNS.

To use Vault, I added a script `generate-and-apply-vault` that 
works like the various `generate-and-apply-*` scripts in the 
TKG lab repository. It uses the Vault Helm chart from Hashicorp
and fills in values from `PARAMS_YAML`. Vault is deployed 
using `kapp` just like other lab components. There's a manual
step to log in and initialize vault the first time through,
but after that it's automatically unsealed using Google.

For Concourse to interact with Vault, I added a YTT overlay to
apply to the Concourse Helm manifest after running 
`genereate-and-apply-concourse`. A script `add-concourse-to-vault`
does some vault configuration then patches and deploys the 
Concoures manifests.

Minio is added with a similar `genereate-and-apply-minio` script
and pre-configured with a bucket and user account for Harbor. 
The pre-configuration uses the same values from `PARAMS_YAML` 
as the Harbor configuration that connects it to an S3-ish 
store to help minimize errors.

Changes
-------

I've diverted from the TKG Lab path by not using any VMware SaaS
products as part of my implementation. I'm using a vanilla Valero
backed up to GCS instead of backups through TMC, and I leverage 
the code that TKG Lab had before it used TMC for backups. Ironically
I made that change when I was at VMW.

I don't have replacements for Wavefont and TMC Policies at this
point, but I do plan on adding observability sometime soon. I 
may or may not do anything with policy, though I suspect I will
add a plain old gatekeeper implementation and some of the policies
from my Governance Lab.