Fixes Concourse connection to Vault
TL;DR
-----

Repairs a defect where Concourse wasn't connecting to Vault

Details
-------

Updates the Concourse/Vault overlay and driver script to address a
defect where the connection was timing out. It was failing because
the credentials were in the wrong format, and when reapplying them
I found a defect in the script. Both are fixed here.