Simplifies setup of minio users
TL;DR
-----

Implements a script to add users to minio

Details
-------

Adds a scriupt for creating a new Minio users and associating a
policy with them, along with a policy for a `platform-team` user
that Concourse can use to connect when doing platform things.

Uses a convention where the policy for the user is loaded from
`${PROJECT_DIR}/minio/policy/${USER}.json` and named for the user.