Adds script for creating a k8s service account (#12)
TL;DR
-----

Implement a helper script to create a service acccount

Details
-------

Supports easily creating a Kuberentes service account to use for
connecting to the cluster. I needed this to support access from
Concourse to deploying stuff, and after stumbling through it a
couple of times manually I automated away the pain.