Repairs Velero backups
TL;DR
-----

Updates Velero install to get backups working correctly

Details
-------

I noticed Velero backups weren't working and let it sit for a bit
since I was a bit cocky about needing them. Two power issues in the
lab led to the need to rebuild it twice. Suddenly the backup issue
felt a lot more urgent.

This change deploys Velero in a way that properly backs up my lab
to GCP, including PVCs.

* Switches from using vSphere snapshots to restic since the config
  seemed a lot simpler and the limitations on restic aren't likely
  to hold me back too much.
* Adds a weekly backup that backs up all volumes, while the nighly
  only backs up object and logs (since they might be helpful if
  something goes bump in the night).

I wanted to switch to the Velero packages in TCE, but it isn't
quite there yet. I would have had to switch to using S3 to connect
to GCS and as I went down that path it got a bit tricky to make
sure it pointed to the right place. I'm going to drop that on
the backlog but I needed a backup right away since the power
sitatuion in the lab is still unstable and electricians are hard
to come by.