Adds `Brewfile` for installing dependencies
TL;DR
-----

Provides a `Brewfile` to install dependences for this and TCE Lab

Details
-------

Leverages Homebrew's bundling feature to make provide dependencies
for this repository and the TCE Lab repository that it defends on.
This makes it a lot easier to make sure the right things are in
place before getting started.